ide: emulate the A4000 motherboard IDE at $DD2020
The A4000 has the same ATA cable as Gayle behind a different decode, so the
drives, task file, and command engine move to a shared AtaBus; Gayle keeps its
ID, interrupt, and PCMCIA registers, and the A4000 interface is a second
front-end. [ide] now fits drives to either machine.

Task file at $DD2020 with Gayle stride, control block at $DD303A, and an
interrupt status register at $DD3020 that follows INTRQ -- Kickstart polls it
after every INT2 and spins there if nothing answers.
