+++ src/lib.rs
@@ -15,6 +15,7 @@ pub mod a2091;
 pub mod a4091;
 pub mod akiko;
 pub mod amigaos;
+pub mod ata;
 pub mod audio;
 pub mod bus;
 pub mod cache;
@@ -38,6 +39,7 @@ pub mod gary;
 pub mod gayle;
 pub mod gdbstub;
 pub mod harddrive;
+pub mod ide_a4000;
 pub mod inputrec;
 pub mod inputsched;
 pub mod memory;
